feat: extract partial dev-utils
