+++ packages/dev-utils/lib/manifest-parser/index.ts
@@ -1,0 +1,2 @@
+import { ManifestParserImpl } from './impl';
+export const ManifestParser = ManifestParserImpl;
