+++ packages/dev-utils/lib/manifest-parser/type.ts
@@ -1,0 +1,7 @@
+import 'chrome';
+
+export type Manifest = chrome.runtime.ManifestV3;
+
+export interface ManifestParserInterface {
+  convertManifestToString: (manifest: Manifest, env: 'chrome' | 'firefox') => string;
+}
