+++ packages/dev-utils/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "@chrome-extension-boilerplate/tsconfig/utils",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "include": ["index.ts", "lib"]
+}
